// Aiminaabee/src/models/fileUpload.ts
export type SizeArrayType = {
  maxHeight: number;
  maxWidth: number;
};

export enum ImageUploadTypes {
  CREATE_USER_IMAGE = 'CREATE_USER_IMAGE',
  CREATE_USER_AVATAR = 'CREATE_USER_AVATAR',
  CREATE_SHOP_IMAGE = 'CREATE_SHOP_IMAGE',
  CREATE_SHOP_AVATAR = 'CREATE_SHOP_AVATAR',
  CREATE_PRODUCT_IMAGES = 'CREATE_PRODUCT_IMAGES',
  CREATE_PRODUCT_DEFAULT_IMAGE = 'CREATE_PRODUCT_DEFAULT_IMAGE',
  CREATE_PRODUCT_DEFAULT_AVATAR = 'CREATE_PRODUCT_DEFAULT_AVATAR',
}

export enum FileNameType {
  SHOP_PHOTO = 'shopPhoto',
  USER_PHOTO = 'userPhoto',
  PRODUCT_PHTO = 'productPhoto',
}